+++ src/ralgo/ralgo_result.rs
@@ -4,7 +4,7 @@ pub struct RalgoResult {
     pub ralgo_calls: u32,
     pub iterations: u32,
     pub calcfg_calls: u32,
-    pub main_circle_radius: f32,
+    pub main_circle_radius: f64,
     pub circles: Vec<Circle>,
 }
 
@@ -13,7 +13,7 @@ impl RalgoResult {
         ralgo_calls: u32,
         iterations: u32,
         calcfg_calls: u32,
-        main_circle_radius: f32,
+        main_circle_radius: f64,
         circles: Vec<Circle>,
     ) -> Self {
         RalgoResult {
